remane test_output_format as test_helper
